requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
filelock==3.14.0
Flask==3.0.3
Flask-WTF==1.2.1
fsspec==2024.3.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
sympy==1.12
tbb==2021.12.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
Werkzeug==3.0.3
WTForms==3.1.2
